backend/requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
google-genai==1.41.0
python-dotenv==1.0.1